+++ src/aux-common/Files/FilesChannel.spec.ts
@@ -1517,7 +1517,7 @@ describe('FilesChannel', () => {
                         id: 'thisFile',
                         tags: {
                             'addItem()':
-                                'applyDiff(@name("bob"), makeDiff.addToMenu())',
+                                'applyDiff(@name("bob"), diff.addToMenu())',
                         },
                     },
                     userFile: {
@@ -1616,7 +1616,7 @@ describe('FilesChannel', () => {
                         id: 'thisFile',
                         tags: {
                             'addItem()':
-                                'applyDiff(@name("bob"), makeDiff.removeFromMenu())',
+                                'applyDiff(@name("bob"), diff.removeFromMenu())',
                         },
                     },
                     userFile: {
@@ -1820,7 +1820,7 @@ describe('FilesChannel', () => {
                         id: 'thisFile',
                         tags: {
                             'test()':
-                                'applyDiff(this, makeDiff.addToContext("abc"))',
+                                'applyDiff(this, diff.addToContext("abc"))',
                         },
                     },
                 };
@@ -1853,7 +1853,7 @@ describe('FilesChannel', () => {
                         tags: {
                             abc: true,
                             'test()':
-                                'applyDiff(this, makeDiff.removeFromContext("abc"))',
+                                'applyDiff(this, diff.removeFromContext("abc"))',
                         },
                     },
                 };
@@ -1949,7 +1949,7 @@ describe('FilesChannel', () => {
                         id: 'thisFile',
                         tags: {
                             'test()':
-                                'applyDiff(this, makeDiff.setPosition("abc", 1, 2))',
+                                'applyDiff(this, diff.setPosition("abc", 1, 2))',
                         },
                     },
                 };
@@ -1977,7 +1977,7 @@ describe('FilesChannel', () => {
                         id: 'thisFile',
                         tags: {
                             'test()':
-                                'applyDiff(this, makeDiff.setPosition("abc", undefined, 2))',
+                                'applyDiff(this, diff.setPosition("abc", undefined, 2))',
                         },
                     },
                 };
@@ -2004,7 +2004,7 @@ describe('FilesChannel', () => {
                         id: 'thisFile',
                         tags: {
                             'test()':
-                                'applyDiff(this, makeDiff.setPosition("abc", undefined, undefined, 2))',
+                                'applyDiff(this, diff.setPosition("abc", undefined, undefined, 2))',
                         },
                     },
                 };