+++ CHANGELOG.md
@@ -10,7 +10,7 @@
     -   Set builder's default background color to dark gray. Player remains the light blue.
     -   Changed the `onDragAny/onDropAny` actions to be `onAnyDrag/onAnyDrop`.
     -   `formula-lib.ts` has changed `isPlayerInContext` export to `player.isInContext`.
--   Bug Fixes
+    -   `formula-lib.ts` has changed `makeDiff` export to `diff`.
 
 ## V0.5.1